% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{clean the RAW data, output clean dataframe}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{Raw data need to be cleaned}
}
\description{
clean the RAW data, output clean dataframe
}
\examples{
\dontrun{
cleaned <- eq_clean_data(raw_data)
}
}
