% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{This functions use to plot earthquakes on relavent map.}
\usage{
eq_map(map_dat, annot_col)
}
\arguments{
\item{annot_col}{columns used to annotate}
}
\description{
This functions use to plot earthquakes on relavent map.
}
\examples{
\dontrun{
labeled_map_dat <- map_dat \%>\% mutate(pop_text = eq_create_label(.))
labeled_map_dat \%>\% eq_map(DATE)
}
}
