% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{plot Timeline without labels}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data used to plot}

\item{stat}{calculation that used to transform the data}

\item{position}{relative position of the geoms}

\item{na.rm}{NA remove}

\item{show.legend}{whether show legend}

\item{inherit.aes}{aes that inherit}

\item{...}{other params to pass in}
}
\description{
plot Timeline without labels
}
\details{
x A date data that used to draw the timeline
  y denote each country
  size same as in ggplot2, used to show magnitude of earth quake
}
\examples{
\dontrun{
ggplot(plot_dat, aes(x = DATE,y = COUNTRY))+ geom_timeline(aes(fill = DEATHS, size = EQ_MAG_MS))
}
}
