% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Geom obj: Timeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\arguments{
\item{x}{A date data that used to draw the timeline}

\item{y}{denote each country}

\item{size}{same as in ggplot2, used to show magnitude of earth quake}
}
\value{
This function use along ggplot2 to draw timeline geom
}
\description{
Geom obj: Timeline
}
\keyword{datasets}
