% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\docType{data}
\name{noaa}
\alias{noaa}
\title{earthquake dataset use for vignette.}
\format{A dataframe with 49 variables and 5998 obs.}
\source{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
noaa
}
\description{
earthquake dataset use for vignette.
}
\keyword{datasets}
