% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timelinelabel}
\alias{geom_timelinelabel}
\title{plot Timeline with labels}
\usage{
geom_timelinelabel(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
plot Timeline with labels
}
\examples{
\dontrun{
ggplot(plot_dat, aes(x = DATE,y = COUNTRY)) + geom_timelinelabel(aes(fill = DEATHS, size = EQ_MAG_MS, loc = LOCATION_NAME_))
}
}
