% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelinelabel}
\alias{GeomTimelinelabel}
\title{Geom Object:Timeline with labels}
\format{An object of class \code{GeomTimelinelabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimelinelabel
}
\arguments{
\item{x}{A date data that used to draw the timeline}

\item{loc}{location of the earthquake, used in the labels}

\item{n_max}{show how many largest earthquake you want show}

\item{y}{denote each country}

\item{size}{same as in ggplot2, used to show magnitude of earth quake}
}
\value{
This function use along ggplot2 to draw timeline geom
}
\description{
Geom Object:Timeline with labels
}
\keyword{datasets}
